:art: chore(remove-duplicate-module): remove duplicate mock-gcal modules
