+++ packages/backend/src/__tests__/mocks.gcal/factories/gcal.event.factory.test.ts
@@ -2,7 +2,7 @@ import { gSchema$EventInstance } from "@core/types/gcal";
 import {
   mockRecurringGcalBaseEvent,
   mockRecurringGcalInstances,
-} from "./gcal.event.factory";
+} from "@backend/__tests__/mocks.gcal/factories/gcal.event.factory";
 
 describe("mockRecurringInstances", () => {
   it("should not include 'recurrence'", () => {
@@ -12,11 +12,13 @@ describe("mockRecurringInstances", () => {
       expect(instance).not.toHaveProperty("recurrence");
     });
   });
+
   it("should create the correct number of instances", () => {
     const base = mockRecurringGcalBaseEvent();
     const instances = mockRecurringGcalInstances(base, 2, 7);
     expect(instances).toHaveLength(2);
   });
+
   it("should include 'recurringEventId' that points to the base event", () => {
     const base = mockRecurringGcalBaseEvent();
     const instances = mockRecurringGcalInstances(base, 2, 7);
@@ -24,6 +26,7 @@ describe("mockRecurringInstances", () => {
       expect(instance.recurringEventId).toBe(base.id);
     });
   });
+
   it("should make the first instance start and end at the same time as the base event", () => {
     const base = mockRecurringGcalBaseEvent();
     const instances = mockRecurringGcalInstances(base, 2, 7);
@@ -32,6 +35,7 @@ describe("mockRecurringInstances", () => {
     expect(firstInstance.start?.dateTime).toBe(base.start?.dateTime);
     expect(firstInstance.end?.dateTime).toBe(base.end?.dateTime);
   });
+
   it("should make instances start and end in the future from the base time (except for the first one)", () => {
     const base = mockRecurringGcalBaseEvent();
     const instances = mockRecurringGcalInstances(base, 2, 7);
@@ -44,6 +48,27 @@ describe("mockRecurringInstances", () => {
       const instanceEnd = new Date(instance.end?.dateTime as string);
       const baseEnd = new Date(base.end?.dateTime as string);
       expect(instanceEnd.getTime()).toBeGreaterThan(baseEnd.getTime());
+    });
+  });
+
+  it("should create recurring instances", () => {
+    const event = mockRecurringGcalBaseEvent();
+    const instances = mockRecurringGcalInstances(event, 3, 7);
+    expect(instances).toHaveLength(3);
+  });
+
+  it("should use RFC3339_OFFSET for start and end times", () => {
+    const event = mockRecurringGcalBaseEvent();
+    const instances = mockRecurringGcalInstances(event, 3, 7);
+    const hasTZOffset = (ts: string) => {
+      return (
+        // @ts-expect-error assuming string has enough length
+        ts[ts.length - 3] === ":" && ["+", "-"].includes(ts[ts.length - 6])
+      );
+    };
+    instances.forEach((instance) => {
+      expect(hasTZOffset(instance.start?.dateTime as string)).toBe(true);
+      expect(hasTZOffset(instance.end?.dateTime as string)).toBe(true);
     });
   });
 });
