--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/delete/all/index.ts
@@ -1,4 +1,0 @@
-import { deleteAllPayload1 } from "./all-1";
-import { deleteAllPayload2 } from "./all-2";
-
-export const deleteAllPayloads = [deleteAllPayload1, deleteAllPayload2];
