--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/delete/single/index.ts
@@ -1,7 +1,0 @@
-import { deleteSinglePayload } from "./delete.single";
-import { deleteSinglePayload2 } from "./delete.single-2";
-
-export const deleteSingleEventPayloads = [
-  deleteSinglePayload,
-  deleteSinglePayload2,
-];
