--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/delete/this-and-following/index.ts
@@ -1,7 +1,0 @@
-import { deleteThisAndFollowingPayload } from "./delete.this-and-following";
-import { deleteThisAndFollowingPayload2 } from "./delete.this-and-following-2";
-
-export const deleteThisAndFollowingPayloads = [
-  deleteThisAndFollowingPayload,
-  deleteThisAndFollowingPayload2,
-];
