--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/edit/all/index.ts
@@ -1,4 +1,0 @@
-import { allPayload1 } from "./all-1";
-import { allPayload2 } from "./all-2";
-
-export const editAllPayloads = [allPayload1, allPayload2];
