--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/edit/single/index.ts
@@ -1,11 +1,0 @@
-import { singleInstance1Payload } from "./single.instance-1";
-import { singleInstance2Payload } from "./single.instance-2";
-import { singleInstance3Payload } from "./single.instance-3";
-import { singleInstance4Payload } from "./single.instance-4";
-
-export const editSingleEventPayloads = [
-  singleInstance1Payload,
-  singleInstance2Payload,
-  singleInstance3Payload,
-  singleInstance4Payload,
-];
