--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/edit/this-and-following/batch1/index.ts
@@ -1,9 +1,0 @@
-import { thisAndFollowing0Payload } from "./this-and-following-v2-0";
-import { thisAndFollowing1Payload } from "./this-and-following-v2-1";
-import { thisAndFollowing2Payload } from "./this-and-following-v2-2";
-
-export const thisAndFollowingBatch1Payloads = [
-  thisAndFollowing0Payload,
-  thisAndFollowing1Payload,
-  thisAndFollowing2Payload,
-];
