--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/edit/this-and-following/batch2/index.ts
@@ -1,11 +1,0 @@
-import { thisAndFollowing1Payload } from "./this-and-following-1";
-import { thisAndFollowing2Payload } from "./this-and-following-2";
-import { thisAndFollowing3Payload } from "./this-and-following-3";
-import { thisAndFollowing4Payload } from "./this-and-following-4";
-
-export const thisAndFollowingBatch2Payloads = [
-  thisAndFollowing1Payload,
-  thisAndFollowing2Payload,
-  thisAndFollowing3Payload,
-  thisAndFollowing4Payload,
-];
