--- packages/backend/src/__tests__/mocks.gcal/mocks.gcal/fixtures/recurring/edit/this-and-following/index.ts
@@ -1,7 +1,0 @@
-import * as batch1 from "./batch1";
-import * as batch2 from "./batch2";
-
-export const thisAndFollowingPayloads = [
-  ...batch1.thisAndFollowingBatch1Payloads,
-  ...batch2.thisAndFollowingBatch2Payloads,
-];
